Testing --orphan checkout.
Checkout into a new branch with the orphan option will guarantee the
first commit made on the new branch will have no parents and will be the
root of a new history totally disconnected from all the other branches
and commits. I think this is what Mike was asking about.